Fix examples to work with most recent release.
The CMake version required to build dd-opentracing-cpp was increased
recently, but the examples/ were not updated to use a version of
Ubuntu that had the required version.

This commit updates the Docker base images used in the examples,
as well as tweaks the examples to build faster and a couple of other
conveniences.